[ADD] 아바타 이동 및 화면 중앙 넘어갈 시 배경도 따라 이동하도록 구현
