+++ front-end/src/App.css
@@ -1,3 +1,5 @@
-body {
+.App {
+  width: 100vw;
+  height: 100vh;
   overflow: hidden;
 }
