+++ front-end/src/components/Town/TownMap/Map/Avatar/style.scss
