+++ front-end/src/components/Town/TownMap/Map/style.scss
@@ -2,4 +2,9 @@
   height: 100%;
   width: 100%;
   position: relative;
+  overflow: hidden;
+
+  & > #map-canvas {
+    position: absolute;
+  }
 }
