+++ front-end/src/components/Town/TownMap/index.js
@@ -1,11 +1,9 @@
 import Map from './Map';
-import Avatar from './Avatar';
 
 const TownMap = () => {
   return (
     <>
       <Map />
-      <Avatar />
     </>
   )
 }
