+++ front-end/src/components/Town/style.scss
@@ -1,0 +1,6 @@
+.town-map {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
